py/requirements.txt
tornado
python-dotenv
dropbox
pypco
pytz
-e git+https://github.com/billdeitrick/pypco.git#egg=pypco
